requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-Limiter==3.5.1
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
grpcio>=1.48.0
grpcio-tools>=1.48.0
protobuf>=3.20.0
gunicorn
